requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
google-cloud-translate>=3.12.0
